Basic Implementation in EDA.py
