+++ src/preprocess/utils.py
@@ -1,0 +1,4 @@
+import os
+
+
+ROOT = os.getcwd()
